refactor chat
