--- frontend/src/components/ChatList/index.js
@@ -1,1 +1,0 @@
-export { default } from './ChatList'
