--- frontend/src/components/Message/Message.module.css
@@ -1,27 +1,0 @@
-.message {
-  display: flex;
-  align-items: flex-start;
-  max-width: 80%;
-  margin-right: auto;
-}
-
-.content {
-  display: flex;
-  flex-direction: column;
-  padding: 20px;
-  margin-bottom: 20px;
-  border-radius: 5px;
-  box-shadow: 0 0 5px 5px rgba(0, 0, 0, 0.2);
-  background-color: #283593;
-  overflow: hidden;
-}
-
-.mla {
-  flex-direction: row-reverse;
-  margin-right: 0;
-  margin-left: auto;
-}
-
-.h3 {
-  margin-bottom: 10px;
-}
