--- frontend/src/components/Message/index.js
@@ -1,1 +1,0 @@
-export { default } from './Message'
