--- frontend/src/components/MessageList/index.js
@@ -1,1 +1,0 @@
-export { default } from './MessageList'
